Major Update: KV cache fixes, ASI cleanup, compiler warnings, format fixes
- PagedKVCache1T: TABLE_ENTRIES 256->4096 for proper 1T capacity
- kv_cache.cpp: auto-create disk directories on offload, fix eviction order in load_from_disk
- format_planner.cpp: fix int64_t loop var to avoid truncation warning
- asi.cpp: remove unused escape_path, suppress blank_lines warning
- trainer.cpp: remove unused quantize_params_to_fp16
- compiler.cmake: add comprehensive Clang warning suppressions
- CMakeLists.txt: add test_asi_flywheel target
- paged_kv_1t_test: use TABLE_ENTRIES constant instead of magic 256

+++ src/format_planner.cpp
@@ -46,7 +46,7 @@ void FormatPlanner::score_importance_fisher(const Tensor& weights,
         double score = 0;
         int64_t start = b * block_size;
         int64_t end = (std::min)(start + block_size, n);
-        for (int i = start; i < end; i++) {
+        for (int64_t i = start; i < end; i++) {
             float g = gd[i];
             score += (double)(g * g);
         }
